requirements.txt
Flask
tensorflow
Pillow
opencv-python-headless
redis
python-dotenv
google-generativeai
gtts
reportlab
numpy
gunicorn
requests
